streamlit_app/requirement.txt
streamlit==1.41.1
pillow==11.0.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2